+++ build-config/fragments/common.js
@@ -6,23 +6,35 @@
 const path = require('path');
 const webpack = require('webpack');
 const moment = require('moment');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+const ReplaceInFileWebpackPlugin = require('replace-in-file-webpack-plugin');
 
 const time = moment().format('YYYY');
 const pckg = require(path.join(__dirname, '..', '..', 'package.json'));
 
 // add banner with copyright and version info
-var bannerPlugin = new webpack.BannerPlugin(
-`${pckg.name}
+var banner = `${pckg.name}
 @version ${pckg.version}
 @see ${pckg.homepage}
 @copyright 2014-${time} ${pckg.author}
-@license ${pckg.license}`
-);
+@license ${pckg.license}`;
+var bannerPlugin = new webpack.BannerPlugin({
+    banner: banner,
+    test: /\.js$/
+});
 // inject version number
 var replaceVersionPlugin = new webpack.DefinePlugin({
   '__VERSION__': JSON.stringify(pckg.version)
 });
-let rootDir = path.resolve(__dirname, '..', '..');
+var cssVersionPlugin = new ReplaceInFileWebpackPlugin([{
+    dir: 'dist/css',
+    test: [/\.css$/],
+    rules: [{
+        search: /LIB_VERSION/ig,
+        replace: pckg.version
+    }]
+}]);
+var rootDir = path.resolve(__dirname, '..', '..');
 
 module.exports = {
     context: rootDir,
@@ -49,11 +61,21 @@ module.exports = {
                 use: {
                     loader: 'babel-loader'
                 }
+            },
+            {
+                test: /\.scss$/,
+                include: path.resolve(rootDir, 'src', 'css'),
+                use: [
+                    MiniCssExtractPlugin.loader,
+                    'css-loader',
+                    'sass-loader'
+                ]
             }
         ]
     },
     plugins: [
         bannerPlugin,
-        replaceVersionPlugin
+        replaceVersionPlugin,
+        cssVersionPlugin
     ]
 };