+++ build-config/fragments/dev.js
@@ -4,6 +4,7 @@
  */
 
 const path = require('path');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 
 module.exports = {
     mode: 'development',
@@ -12,5 +13,10 @@ module.exports = {
         contentBase: [path.resolve(__dirname, '..', '..')],
         publicPath: 'localhost:8080/dist/',
         watchContentBase: true
-    }
+    },
+    plugins: [
+        new MiniCssExtractPlugin({
+            filename: 'css/[name].css'
+        })
+    ]
 };